:tractor: Split filewalker
